+++ R/load.R
@@ -28,10 +28,9 @@ loadStudyData <- function(tz, override = NULL, start.time, stop.time, save.detec
 
   # Check that all the overridden fish are part of the study
   if (!is.null(override)) {
-    override_signals <- unlist(lapply(strsplit(override, "-"), function(x) tail(x, 1)))
     lowest_signals <- sapply(bio$Signal, function(i) min(as.numeric(unlist(strsplit(as.character(i), "|", fixed = TRUE)))))
-    if (any(link <- is.na(match(override_signals, lowest_signals))))
-      stop("Some tag signals listed in 'override' ('", paste0(override[link], collapse = "', '"), "') are not listed in the biometrics file.\n", call. = FALSE)
+    if (any(link <- is.na(match(override, lowest_signals))))
+      stopAndReport("Some tag signals listed in 'override' ('", paste0(override[link], collapse = "', '"), "') are not listed in the biometrics file.")
   }
   deployments <- loadDeployments(input = "deployments.csv", tz = tz)
   checkDeploymentTimes(input = deployments) # check that receivers are not deployed before being retrieved