+++ R/migration.R
@@ -222,7 +222,7 @@ by which sections are presented", immediate. = TRUE, call. = FALSE)
     ", minimum.detections = ", minimum.detections,
     ", start.time = ", ifelse(is.null(start.time), "NULL", paste0("'", start.time, "'")),
     ", stop.time = ", ifelse(is.null(stop.time), "NULL", paste0("'", stop.time, "'")),
-    ", speed.method = ", paste0("c('", speed.method, "')"),
+    ", speed.method = '", speed.method, "'",
     ", speed.warning = ", ifelse(is.null(speed.warning), "NULL", speed.warning),
     ", speed.error = ", ifelse(is.null(speed.error), "NULL", speed.error),
     ", jump.warning = ", jump.warning,
@@ -230,7 +230,7 @@ by which sections are presented", immediate. = TRUE, call. = FALSE)
     ", inactive.warning = ", ifelse(is.null(inactive.warning), "NULL", inactive.warning),
     ", inactive.error = ", ifelse(is.null(inactive.error), "NULL", inactive.error),
     ", exclude.tags = ", ifelse(is.null(exclude.tags), "NULL", paste0("c('", paste(exclude.tags, collapse = "', '"), "')")),
-    ", override = ", ifelse(is.null(override), "NULL", paste0("c('", paste(override, collapse = "', '"), "')")),
+    ", override = ", ifelse(is.null(override), "NULL", paste0("c(", paste(override, collapse = ", "), ")")),
     ", report = ", ifelse(report, "TRUE", "FALSE"),
     ", auto.open = ", ifelse(auto.open, "TRUE", "FALSE"),
     ", discard.orphans = ", ifelse(discard.orphans, "TRUE", "FALSE"),
@@ -240,7 +240,7 @@ by which sections are presented", immediate. = TRUE, call. = FALSE)
     ", disregard.parallels = ", ifelse(disregard.parallels, "TRUE", "FALSE"),
     ", GUI = '", GUI, "'",
     ", print.releases = ", ifelse(print.releases, "TRUE", "FALSE"),
-    ", plot.detections.by = ", plot.detections.by,
+    ", plot.detections.by = '", plot.detections.by, "'",
     ")")
 
   appendTo("debug", the.function.call)
@@ -376,7 +376,7 @@ by which sections are presented", immediate. = TRUE, call. = FALSE)
 
   movement.names <- names(movements)
 
-  if (any(link <- !override %in% movement.names)) {
+  if (any(link <- !override %in% extractSignals(movement.names))) {
     appendTo(c("Screen", "Warning", "Report"), paste0("Override has been triggered for fish ", paste(override[link], collapse = ", "), " but ",
       ifelse(sum(link) == 1, "this", "these"), " fish ", ifelse(sum(link) == 1, "was", "were")," not detected."))
     override <- override[!link]
@@ -386,7 +386,7 @@ by which sections are presented", immediate. = TRUE, call. = FALSE)
     fish <- names(movements)[i]
     appendTo("debug", paste0("debug: Checking movement quality for fish ", fish,"."))
 
-    if (is.na(match(fish, override))) {
+    if (is.na(match(extractSignals(fish), override))) {
       release <- as.character(bio$Release.site[na.as.false(bio$Transmitter == fish)])
       release <- unlist(strsplit(with(spatial, release.sites[release.sites$Standard.name == release, "Array"]), "|", fixed = TRUE))